+++ lib/rules/no-broken-template.js
@@ -3,8 +3,10 @@
  * @author Eugene Zhlobo
  */
 
-const { findVariablesInTemplate } = require('pug-uses-variables')
-const { isReactPugReference, getTemplate, buildLocation } = require('../utilities')
+const lexer = require('pug-lexer')
+
+const { isReactPugReference, buildLocation } = require('../util/eslint')
+const getTemplate = require('../util/getTemplate')
 
 //------------------------------------------------------------------------------
 // Rule Definition
@@ -25,9 +27,8 @@ module.exports = {
       TaggedTemplateExpression: function (node) {
         if (isReactPugReference(node)) {
           const template = getTemplate(node)
-
           try {
-            findVariablesInTemplate(template)
+            lexer(template)
           } catch (error) {
             const line = (error.line + node.loc.start.line) - 1
             const source = error.src.split('\n')[error.line - 1]
