+++ lib/rules/no-undef.js
@@ -3,12 +3,9 @@
  * @author Eugene Zhlobo
  */
 
-const {
-  isReactPugReference,
-  getTemplate,
-  extractVariables,
-  buildLocation,
-} = require('../utilities')
+const { isReactPugReference, buildLocation } = require('../util/eslint')
+const getTemplate = require('../util/getTemplate')
+const getVariables = require('../util/getVariables')
 
 //------------------------------------------------------------------------------
 // Rule Definition
@@ -42,7 +39,7 @@ module.exports = {
         if (isReactPugReference(node)) {
           const template = getTemplate(node)
 
-          const usedVariables = extractVariables(template)
+          const usedVariables = getVariables(template)
           const definedVariables = getDefinedVariables()
 
           const isVariableDefined = variable => !definedVariables.includes(variable.value)
