+++ lib/rules/quotes.js
@@ -3,10 +3,12 @@
  * @author Eugene Zhlobo
  */
 
-const { isReactPugReference, getTemplate, buildLocation } = require('../utilities')
-const lexer = require('pug-lexer')
 const { parse } = require('@babel/parser')
 const codeWalk = require('@babel/traverse').default
+
+const { isReactPugReference, buildLocation } = require('../util/eslint')
+const getTemplate = require('../util/getTemplate')
+const getTokens = require('../util/getTokens')
 
 //------------------------------------------------------------------------------
 // Rule Definition
@@ -43,7 +45,7 @@ module.exports = {
           const withWrongQuotes = []
 
           const template = getTemplate(node)
-          const tokens = lexer(template)
+          const tokens = getTokens(template)
 
           tokens.forEach((token) => {
             const normalizedValue = normalizeValue(token)
