+++ lib/rules/uses-react.js
@@ -3,7 +3,7 @@
  * @author Eugene Zhlobo
  */
 
-const { isReactPugReference } = require('../utilities')
+const { isReactPugReference } = require('../util/eslint')
 
 //------------------------------------------------------------------------------
 // Rule Definition
