+++ lib/rules/uses-vars.js
@@ -3,7 +3,9 @@
  * @author Eugene Zhlobo
  */
 
-const { isReactPugReference, getTemplate, extractVariables } = require('../utilities')
+const { isReactPugReference } = require('../util/eslint')
+const getTemplate = require('../util/getTemplate')
+const getVariables = require('../util/getVariables')
 
 //------------------------------------------------------------------------------
 // Rule Definition
@@ -25,7 +27,7 @@ module.exports = {
         if (isReactPugReference(node)) {
           const template = getTemplate(node)
 
-          extractVariables(template)
+          getVariables(template)
             .forEach(variable => context.markVariableAsUsed(variable.value))
         }
       },
