+++ lib/util/getTemplate.js
@@ -1,5 +1,3 @@
-const { findVariablesInTemplate } = require('pug-uses-variables')
-
 const PLACEHOLDER = '__eslint-react-pug__'
 
 const getQuasiValue = ({ value }) => {
@@ -19,11 +17,7 @@ const getInterpolatedTemplate = (template, interpolations) => template
   })
   .join('')
 
-function isReactPugReference({ tag }) {
-  return tag && tag.name === 'pug'
-}
-
-function getTemplate({ quasi }) {
+module.exports = ({ quasi }) => {
   const { quasis, expressions } = quasi
 
   if (expressions && expressions.length) {
@@ -31,32 +25,4 @@ function getTemplate({ quasi }) {
   }
 
   return getQuasiValue(quasis[0])
-}
-
-function extractVariables(template) {
-  try {
-    return findVariablesInTemplate(template)
-  } catch (e) {
-    return []
-  }
-}
-
-function buildLocation(startCoord, endCoord) {
-  return {
-    start: {
-      line: startCoord[0],
-      column: startCoord[1],
-    },
-    end: {
-      line: endCoord[0],
-      column: endCoord[1],
-    },
-  }
-}
-
-module.exports = {
-  isReactPugReference,
-  getTemplate,
-  extractVariables,
-  buildLocation,
 }
