+++ lib/util/getTokens.js
@@ -1,0 +1,9 @@
+const lexer = require('pug-lexer')
+
+module.exports = (template) => {
+  try {
+    return lexer(template)
+  } catch (error) {
+    return []
+  }
+}
