+++ lib/util/getVariables.js
@@ -1,0 +1,9 @@
+const { findVariablesInTemplate } = require('pug-uses-variables')
+
+module.exports = (template) => {
+  try {
+    return findVariablesInTemplate(template)
+  } catch (error) {
+    return []
+  }
+}
