+++ package.json
@@ -6,7 +6,7 @@
   "main": "index.js",
   "scripts": {
     "lint": "eslint ./",
-    "test": "mocha tests --recursive"
+    "test": "mocha tests"
   },
   "files": [
     "LICENSE",
